Agregado mas Edicion de Perfil de profesional

+++ app/Http/Controllers/GraduateController.php
@@ -6,9 +6,15 @@ use App\Models\Career;
 use App\Models\Degree;
 use App\Models\Graduate;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class GraduateController extends Controller
 {
+    function __construct()
+    {
+        $this->middleware('auth')->only(['destroy']);
+    }
+
     function index(Request $request)
     {
         $graduates = Graduate::orderBy('id')->where('published', '=', true)->paginate(24);
@@ -20,20 +26,23 @@ class GraduateController extends Controller
         $graduates = $career->graduates()->where('published', '=', true)->paginate(24);
         return view('portal.graduate.career', ['career' => $career, 'graduates' => $graduates]);
     }
-    function create()
-    {
-        return "FORM CREATE HERE";
-    }
+
+
+
     function destroy(Graduate $graduate)
     {
-        $success = $graduate->delete();
+        $user = Auth::user();
+        if (($user->can('graduates.delete') || $user->id == $graduate->person->user_id)) {
+            $success = $graduate->delete();
 
-        if ($success) {
-            return redirect()->back()->with("message", "Se ha borrado correctamente.")
-                ->with("type", "success");
+            if ($success) {
+                return redirect()->back()->with("message", "Se ha borrado correctamente.")
+                    ->with("type", "success");
+            }
+
+            return redirect()->back()->withInput()->with("message", "Algo ha salido mal, vuelva a intentar mas tarde.")
+                ->with("type", "error");
         }
-
-        return redirect()->back()->withInput()->with("message", "Algo ha salido mal, vuelva a intentar mas tarde.")
-            ->with("type", "error");
+        abort(403, 'No autorizado para realizar esta accion.');
     }
 }
